+++ CHANGELOG.md
@@ -51,6 +51,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
   - Fixed decide sending mailer in `MailWorkerCommand` depending on template mail type.
 - Fixed type error authentication issue when Mailer uses REMP SSO for authentication. remp/remp#1161
 - Fixed internal content parsers which could throw notices if parsed URL doesn't contain JSON schema. remp/remp#1162
+- Added support for allow listing emails in Mailer. remp/remp#1150
+  - Added tests for application `Sender` and `EmailAllowList` classes.
 
 ## [0.32.2] - 2022-05-11